create index.html
Lab 1 steps:
Create a new file called index.html in the root of your repository
Type ! and hit Tab to expand the Emmet snippet into a basic HTML page
Change the page title (<title>) to “YOUR_NAME: Personal site and portfolio”
Add a <h1> with your name
Add a <p> with a short description of yourself (if you can’t think of anything, just type lorem and hit Tab to generate placeholder text; just make sure to edit it later!)
Save the file. Open the file in Chrome and verify that it works.
Commit and push your changes to GitHub